added encode and rename file
